requirements.txt
absl-py==0.15.0
altair==4.2.2
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.9.3
bleach==6.1.0
blinker==1.7.0
brewer2mpl==1.4.1
cachelib==0.12.0
cachetools==4.2.4
catboost==1.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
coolname==1.1.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
deepface==0.0.49
defusedxml==0.7.1
dlib==19.24.2
dnspython==2.6.1
email-validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-Session==0.3.2
Flask-WTF==1.0.0
flatbuffers==1.12
gast==0.3.3
gdown==4.5.1
gevent==24.2.1
ggplot==0.6.6
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.0.1
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.32.0
gunicorn==20.1.0
h11==0.14.0
h5py==2.10.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
importlib-metadata==7.1.0
importlib-resources==6.4.0
imutils==0.5.3
infinity==1.5
intervals==0.9.2
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==3.0.3
joblib==1.3.2
json5==0.9.24
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.1
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter-server==2.13.0
jupyter-server-terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.4
jupyterlab-widgets==3.0.10
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.41.1
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.3.3
matplotlib-inline==0.1.6
mistune==3.0.2
msgspec==0.18.6
mtcnn==0.1.1
mysqlclient==2.2.4
narwhals==1.10.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
nltk==3.6.2
notebook==7.1.2
notebook-shim==0.2.4
np-utils==0.6.0
numba==0.47.0
numpy==1.20.2
oauthlib==3.2.2
object-detection==0.0.3
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.0
pandas==1.2.4
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pickleshare==0.7.5
Pillow==8.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
plotly==5.22.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydeck==0.9.1
pygments==2.17.2
pyparsing==3.1.2
PySimpleGUI==5.0.7
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
python-json-logger==2.0.7
python-whois==0.9.4
pytz==2021.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.25.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.2
six==1.15.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.0.0
stripe==2.27.0
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.2
tensorflow-estimator==2.2.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
terminado==0.18.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tornado==6.4
tqdm==4.42.1
traitlets==5.14.2
tslib==1.6
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.24.3
validators==0.27.0
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==1.0.1
wget==3.2
whois==0.9.13
widgetsnbextension==4.0.10
wrapt==1.12.1
WTForms==2.1
WTForms-Components==0.10.5
zipp==3.18.1
zope.event==5.0
zope.interface==7.0.3